+++ test/optionsResolverTests.js
@@ -431,12 +431,13 @@ exports.tests = {
     },
     "Visual Studio `vs` Integration Tests": {
         "Visual Studio properties come across as expected": function (test) {
-            test.expect(2);
+            test.expect(3);
             var cfg = getConfig("vs minimalist");
             cfg.options = { sourceMap: false };
             var result = or.resolveAsync(null, cfg).then(function (result) {
                 test.strictEqual(result.allowSyntheticDefaultImports, true);
                 test.strictEqual(result.noStrictGenericChecks, true);
+                test.strictEqual(result.esModuleInterop, true);
                 test.done();
             }).catch(function (err) { test.ifError(err); test.done(); });
         },
@@ -701,7 +702,7 @@ exports.tests = {
             }).catch(function (err) { test.ifError(err); test.done(); });
         },
         "config entries come through appropriately": function (test) {
-            test.expect(23);
+            test.expect(24);
             var cfg = getConfig("minimalist");
             cfg.tsconfig = './test/tsconfig/full_valid_tsconfig.json';
             var result = or.resolveAsync(null, cfg).then(function (result) {
@@ -722,6 +723,7 @@ exports.tests = {
                 test.strictEqual(result.checkJs, true);
                 test.strictEqual(result.allowJs, true);
                 test.strictEqual(result.preserveSymlinks, true);
+                test.strictEqual(result.esModuleInterop, true);
                 test.strictEqual(result.types.length, 3);
                 test.ok(result.types.indexOf('issue') > -1);
                 test.strictEqual(result.typeRoots.length, 2);