+++ test/optionsResolverTests.ts
@@ -449,12 +449,13 @@ export var tests : nodeunit.ITestGroup = {
 
   "Visual Studio `vs` Integration Tests": {
     "Visual Studio properties come across as expected": (test: nodeunit.Test) => {
-      test.expect(2);
+      test.expect(3);
       const cfg = getConfig("vs minimalist");
       cfg.options = <any>{sourceMap : false};
       const result = or.resolveAsync(null, cfg).then((result) => {
         test.strictEqual(result.allowSyntheticDefaultImports, true)
         test.strictEqual(result.noStrictGenericChecks, true)
+        test.strictEqual(result.esModuleInterop, true)
         test.done();
       }).catch((err) => {test.ifError(err); test.done();});
     },
@@ -734,7 +735,7 @@ export var tests : nodeunit.ITestGroup = {
         }).catch((err) => {test.ifError(err); test.done();});
     },
     "config entries come through appropriately": (test: nodeunit.Test) => {
-        test.expect(23);
+        test.expect(24);
         const cfg = getConfig("minimalist");
         cfg.tsconfig = './test/tsconfig/full_valid_tsconfig.json';
 
@@ -756,6 +757,7 @@ export var tests : nodeunit.ITestGroup = {
           test.strictEqual(result.checkJs, true);
           test.strictEqual(result.allowJs, true);
           test.strictEqual(result.preserveSymlinks, true);
+          test.strictEqual(result.esModuleInterop, true);
           test.strictEqual(result.types.length, 3);
           test.ok(result.types.indexOf('issue') > -1);
           test.strictEqual(result.typeRoots.length, 2);